Refactor resolver into submodules
